Completely restructured the plugin interface and updated my plugins Changed the magic numbers in the Logitech plugin's MoreIntenseColor() method and removed the unused copy in the main program's Extensions class Removed/disabled the Aero plugin because Windows 10's implementation of DwmSetColorizationParameters apparently changed and I cba to read up on how exactly it did
Added an error handler (which is basically just a catch(all) and a message box) to the ColorManager
Added a few comments in the FocusWatcher
Added list of "forbidden" process IDs to the FocusWatcher
Replaced the plugin detection method in the PluginManager to stop crashes that would occur if a .dll was using an older version of the plugin interface

Made SettingsPluginGUIDIsSelectedConverter not suck (wow, that's a mouthful. i should've been a java developer)

Extended the Settings object to allow custom colors
Also added better error checks to the deserializer

Added try/catch block to the GetEXEIconBitmap helper method because some people *cough*MICROSOFT*cough* apparently put corrupt icon data into their applications

Added custom color support and a reset button
Restyled a few UI components, most notably the plugin list and the hyperlinks throughout the app

Added a new "first start" window. There's nothing on it yet, but the file's there

Last but not least, I bumped the version number to 0.9.